requirements.txt
# local package
-e .

# external requirements
pybids
